requirements.txt
streamlit
pytest
python-dotenv
google-generativeai
